requirements.txt
annotated-types==0.6.0
anyio==3.7.1
boto3==1.28.84
botocore==1.31.84
certifi==2023.7.22
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
idna==3.4
jmespath==1.0.1
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
